@@ -24,8 +24,10 @@ setup(
     packages=find_packages('src'),
     classifiers=[
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
+        'Programming Language :: Python :: 3.12',
         'License :: OSI Approved :: Apache Software License',
         'Operating System :: OS Independent',
         'Topic :: Software Development :: Testing',
@@ -33,11 +35,11 @@ setup(
         'Framework :: Robot Framework',
     ],
     install_requires=[
-        'robotframework >= 3.2.2',
-        'PyNaCl >= 1.4.0',
-        'questionary>=1.9.0',
+        'robotframework >= 6.0.0',
+        'PyNaCl >= 1.5.0',
+        'questionary>=2.0.1',
     ],
-    python_requires='>=3.6',
+    python_requires='>=3.9',
     entry_points={
         'console_scripts': [
             'CryptoLibrary = CryptoLibrary:main',
